record_stream: initial commit
A record stream is a sequence of variable-length records encoded in a
self-synchronised format.  This format makes it easy to populate logs
robustly and concurrently, and lets readers seek in a log file.

Backtrace has been storing critical data using this format for more
than 9 months.  We are standardising on it for all streams of
variable-length records.